# backend/src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TurPoint API",
    "version": "0.3.0",
    "description": "Tour marketplace API - operators, tours, group formation pricing, bookings, reviews, deals, and the smart trip planner. Most write operations and anything scoped to \"the logged-in account\" require a Bearer JWT; identity is always derived from the verified token, never from the request body. Sprint 3: all responses carry standard security headers (helmet); /api/auth/login and /api/auth/signup are rate-limited (20 requests/15min per IP) and every other route shares a broader 300 requests/15min ceiling; in production CORS is restricted to the configured frontend origin via CORS_ORIGIN."
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/operators": {
      "post": {
        "summary": "Create an operator profile - requires login; one profile per account. phone is required and must start with +994. The instagram field is normalised to a bare handle (\"@name\", \"instagram.com/name\" and full URLs are all accepted).",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "name": "Qafqaz Tours",
                "description": "Dag turlari",
                "languages": "az,en",
                "phone": "+994501234567",
                "instagram": "qafqaztours"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Missing name, or missing/invalid phone (must start with +994)"
          },
          "401": {
            "description": "Missing/invalid token"
          },
          "409": {
            "description": "This account already has an operator profile - use PUT to update it instead"
          }
        }
      },
      "get": {
        "summary": "List operators",
        "responses": {
          "200": {
            "description": "Array of operators"
          }
        }
      }
    },
    "/operators/me": {
      "get": {
        "summary": "Get the logged-in account's own operator profile, if any - drives the Traveler/Operator mode toggle on the frontend",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Operator profile, or null if this account isn't an operator yet"
          },
          "401": {
            "description": "Missing/invalid token"
          }
        }
      }
    },
    "/operators/me/photo": {
      "post": {
        "summary": "Upload/replace the logged-in account's operator profile photo - multipart/form-data with a single 'photo' image file field (max 5MB). Requires an existing operator profile (create one first via POST /operators).",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "photo": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated operator, with the new photo_url"
          },
          "400": {
            "description": "No file uploaded, or the file isn't an image"
          },
          "401": {
            "description": "Missing/invalid token"
          },
          "404": {
            "description": "No operator profile yet - create one first"
          }
        }
      }
    },
    "/operators/me/phone/send-code": {
      "post": {
        "summary": "Send a verification code for a phone number, ahead of it being saved. MOCKED - no SMS provider is wired up yet, so this always issues the fixed code 123456 (logged server-side) instead of texting one. Swap in a real provider + random code later without changing the request/response shape.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "phone": "+994501234567"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Code issued (mocked)"
          },
          "400": {
            "description": "Missing/invalid phone (must start with +994)"
          },
          "401": {
            "description": "Missing/invalid token"
          },
          "404": {
            "description": "No operator profile yet - create one first"
          }
        }
      }
    },
    "/operators/me/phone/verify-code": {
      "post": {
        "summary": "Confirm the code from send-code. On success, the pending phone number becomes the operator's actual phone and phone_verified is set to true - this is the only place that happens.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "code": "123456"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Phone verified - updated operator"
          },
          "400": {
            "description": "Missing code, no pending verification, expired code, or incorrect code"
          },
          "401": {
            "description": "Missing/invalid token"
          },
          "404": {
            "description": "No operator profile yet - create one first"
          }
        }
      }
    },
    "/operators/{id}": {
      "get": {
        "summary": "Get one operator",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operator"
          },
          "404": {
            "description": "Not found"
          }
        }
      },
      "put": {
        "summary": "Update an operator profile - must be the owning account. phone is required and must start with +994 (vehicle_features isn't sent by the frontend form anymore, but is preserved untouched if omitted).",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "name": "Qafqaz Tours",
                "description": "Dag turlari",
                "languages": "az,en",
                "phone": "+994501234567",
                "instagram": "qafqaztours"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated operator"
          },
          "400": {
            "description": "Missing/invalid phone (must start with +994)"
          },
          "403": {
            "description": "Not the owning account"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
  }
}